python-pylint: modifies package bin to match source and adds dependencies
The pylint script mismatches from source in the target, producing runtime
errors, such as:
    "pkg_resources.DistributionNotFound: The 'configparser'
     was not found and is required by pylint"
This exception is raised mainly because pkg_resources, which is used by
pylint, looks for modules named as in Py3, but the target has only Py2.
For instance, 'ConfigParser' (already installed), has been renamed to
'configparser' in Py3, and thus not found by pkg_resources.
To ensure none of these compatibility issues happen, the pylint script
handles all the imports.
Thus, /usr/bin/pylint has to be installed in the same fashion as in
source, and the recipe in this patch states the content of such script
and deletes the previous one.

Also, a set of runtime dependencies is added to the recipe.

(From meta-openembedded commit: d9b5870edaddee69612dada2aa4bc2d8c3f70f44)

Signed-off-by: Daniela Plascencia <daniela.plascencia@linux.intel.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>
Signed-off-by: Tim Orling <ticotimo@gmail.com>