python-paho-mqtt: fix downloading during do_compile
During do_compile if setup.py finds that pytest-runner - which is
declared as a setup dependency in setup.py - isn't available it was
attempting to download it from pypi, which is problematic if you are
behind a proxy because the environment for do_compile isn't set up for
that. Add an appropriate native dependency to prevent this situation.

(From meta-openembedded commit: 2951fb898e7c761e3071521b58c9aa6cc746fba3)

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>
Signed-off-by: Tim Orling <ticotimo@gmail.com>