python-jsonschema, python3-jsonschema: add vcversioner-native dependency
If not added the setup.py will complain that it can't find
vcversioner. The non native dependency is not working as the module is
required during compilation not at runtime.

Also the dependency is required for both the Python 2 and Python 3 version

(From meta-openembedded commit: 12d09665bbee2825ed54cd4254a33235d2ca3e26)

Signed-off-by: Pascal Bach <pascal.bach@siemens.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>
Signed-off-by: Tim Orling <ticotimo@gmail.com>