pyinotify: Fix support for different python versions
The runtime dependencies for pyinotify are defined in
python-pyinotify.inc file, which gets included from both the python2 and
the python3 recipes for pyinotify.

Since the include file is used for both python2 and python3 recipes, it
must use the '${PYTHON_PN}-' prefix instead of just 'python-' for
listing the dependencies in RDEPENDS, otherwise it is forcing python2
modules to be selected from a python3 recipe, so fix that.

For example, use '${PYTHON_PN}-threading' instead of 'python-threading'
for specifying a dependency from the 'threading' module.

While at it, also use line separation and indentation on RDEPENDS for
better readability.

(From meta-openembedded commit: a5c7c4746ee6f12f925039ef88a73737e30942b8)

Signed-off-by: Ricardo Silva <rjpdasilva@gmail.com>
Signed-off-by: Armin Kuster <akuster808@gmail.com>
Signed-off-by: Tim Orling <ticotimo@gmail.com>