python-twisted: Add more python-twisted-* packages to the "meta" package
Whereas the python-twisted-core package is used to install the minimum
components of python-twisted, the python-twisted package is
essentially a 'meta' package which installs all of the
python-twisted-* packages (used by folks that don't want to determine
which components they are interested in). Add additional packages to
the python-twisted 'meta' package.

NOTE: that there are still modules of python-twisted which have not
been ported to python3 (see src/twisted/python/_setup.py
"notPortedModules") so the installed packages for py2 vs. py3 will not
be the same (ie. python-twisted-news is only part of the py2 package).

(From meta-openembedded commit: 12c394588cbda922482c14616135030f530e8a2e)

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Tim Orling <ticotimo@gmail.com>