recipes-devtools: merge inc and bb PART I
Since we no longer need to support two python
versions, merge inc and bb.

This makes maintenance easier with AUH and devtool.

While we are at it, drop some dangling inc files
and some python3 recipes.

Enforce Debian naming (python-foo). Only recipes that
are executable tools, such as 'gyp' or 'gateone' should
be exceptions.

When in doubt, search for what the package is named
in Debian or Fedora.

Signed-off-by: Tim Orling <ticotimo@gmail.com>