python-jsonref: merge bb and inc and move
Since we no longer need to support two python
versions, merge inc and bb.

This makes maintenance easier with AUH and devtool.

Add DESCRIPTION.

There is no obvious reason by this recipe needed to be
in its own directory, so move it with the rest of the
recipes in recipes-devel/python.

Signed-off-by: Tim Orling <ticotimo@gmail.com>