kas-meta-python-ptest-image: fixes for GL
To make testing as fast as possible, it is desired
to use KVM and OpenGL.

Add
  DISTRO_FEATURES_append = "opengl"
  PACKAGECONFIG_pn-qemu-system-native = "fdt alsa kvm virglrenderer glx gtk+"

It shouldn't be required to be explicit about the PACKAGECONFIG above,
when all I was trying to do was add gtk+, but but using _pn- += or even
_append_pn- syntax resulted in a system with either disabled OpenGL or
missing canberra-gtk-module (sound). The above explicit definition worked.

Change gl to gl-es:
  TEST_RUNQEMUPARAMS = "kvm gl-es gtk"

Signed-off-by: Tim Orling <ticotimo@gmail.com>